+++ react/src/components/utils/csrf.js
