+++ react/src/pages/pochtron_shop.js
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom'
 import React from 'react'
-import {AddTransaction, LastTransactions} from './components/shop'
+import {AddTransaction, LastTransactions} from '../components/shop'
 
 ReactDOM.render((
     <><AddTransaction></AddTransaction><LastTransactions></LastTransactions></>
